refactor(control-api): extract dto from response_actions/mod.rs
